Remove constraints on the size of the render window
In the original code, the width of the render window is forced to be a multiple of 8, as well as its height to be a multiple of 2. For resolutions like 1366x768, the constraint on the width causes a border to exist on the sides of the screen, since 1366 is not a multiple of 8.

If the complete removal of these constraints may cause some sort of problem, a change only to the width constraint may be considered.