Add a new icon with much higher resoultion (suitable for Win 10).
This icon was supplied by Ryan as SVG. I've converted the raw SVG in a
PNG with size 512x512 (for Linux / Unix desktops) and an ICO (for
Windows). The ICO contains the levels 16x16 to 512x512 and should be
usable on high DPI screens.

The ImageMagick command to generate the ICO was:

convert quake2.png -bordercolor white -border 0  \
    \( -clone 0 -resize 16x16 \) \
    \( -clone 0 -resize 32x32 \) \
    \( -clone 0 -resize 48x48 \) \
    \( -clone 0 -resize 64x64 \) \
    \( -clone 0 -resize 96x96 \) \
    \( -clone 0 -resize 128x128 \) \
    \( -clone 0 -resize 256x256 \) \
    \( -clone 0 -resize 512x512 \) \
    -delete 0 quake2.ico

This closes issue #106.