Add support for .override files (needs testing)
r1q2 has a neat feature: transparent server-side map entity replacement, this is, the loading of external files ending in .override (i.e. base2.bsp.override) which adds or change entities to a map depending on the gamemode on the server.

For example, you want to play CTF on the map base2, but found that the map lacks the flags and spawnpoints, so you use these .override files in order to add CTF support to it by way of adding the flags, spawnpoints and other necessary entities.